+++ src/components/player-header.ts
@@ -72,28 +72,8 @@ class PlayerHeader extends LitElement {
    */
   private styleContainer() {
 
-    // load card configuration theme settings.
-    const playerHeaderTitle1Color = this.config.playerHeaderTitle1Color;
-    const playerHeaderTitle1FontSize = this.config.playerHeaderTitle1FontSize;
-    const playerHeaderTitle2Color = this.config.playerHeaderTitle2Color;
-    const playerHeaderTitle2FontSize = this.config.playerHeaderTitle2FontSize;
-    const playerHeaderTitle3Color = this.config.playerHeaderTitle3Color;
-    const playerHeaderTitle3FontSize = this.config.playerHeaderTitle3FontSize;
-
     // build style info object.
     const styleInfo: StyleInfo = <StyleInfo>{};
-    if (playerHeaderTitle1Color)
-      styleInfo['--stpc-player-header-title1-color'] = `${playerHeaderTitle1Color}`;
-    if (playerHeaderTitle1FontSize)
-      styleInfo['--stpc-player-header-title1-font-size'] = `${playerHeaderTitle1FontSize}`;
-    if (playerHeaderTitle2Color)
-      styleInfo['--stpc-player-header-title2-color'] = `${playerHeaderTitle2Color}`;
-    if (playerHeaderTitle2FontSize)
-      styleInfo['--stpc-player-header-title2-font-size'] = `${playerHeaderTitle2FontSize}`;
-    if (playerHeaderTitle3Color)
-      styleInfo['--stpc-player-header-title3-color'] = `${playerHeaderTitle3Color}`;
-    if (playerHeaderTitle3FontSize)
-      styleInfo['--stpc-player-header-title3-font-size'] = `${playerHeaderTitle3FontSize}`;
     return styleMap(styleInfo);
 
   }
@@ -108,10 +88,10 @@ class PlayerHeader extends LitElement {
       css`
 
       .player-header-container {
-        margin: 0.75rem 3.25rem;
         margin-top: 0rem;
         padding: 0.5rem;
         padding-top: 0rem;
+        padding-bottom: 0rem;
         max-width: 40rem;
         text-align: center;
         background-position: center;