+++ src/utils/media-browser-utils.ts
@@ -104,12 +104,6 @@ export function getContentItemImageUrl(item: any, config: CardConfig, hasItemsWi
  */
 export function getMdiIconImageUrl(mdi_icon: string): string {
 
-  //// assign default color value, if one is not assigned.
-  //if (!mdi_icon_color) {
-  //  mdi_icon_color = "#2196F3"
-  //}
-  //mdi_icon_color = mdi_icon_color.replace("#", "%23");
-
   // assign the icon url.
   const mdiImageUrl = '\'data:image/svg+xml;utf-8,<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path d="' + mdi_icon + '"></path></svg>\'';
   //const mdiImageUrl = '\'data:image/svg+xml;utf-8,<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path fill="' + mdi_icon_color + '" d="' + mdi_icon + '"></path></svg>\'';
@@ -155,6 +149,11 @@ export function formatTitleInfo(
   let result = formatConfigInfo(text, config);
   result = formatPlayerInfo(result, player);
   result = formatMediaListInfo(result, mediaListLastUpdatedOn, mediaList, filteredList);
+
+  // if result is an empty string, then return undefined.
+  if ((result + "").trim() == "")
+    result = undefined;
+
   return result;
 }